Refactor to reflect haskell impl:
- switch to indexmap to maintain order in state maps.
 - switch to 128 bit numbers to match haskell impl.
 - add more tests for datum re-serialization. use str for wasm to support 128 bit apis.
 - drop empty accounts from state.
 - fix tag encoding of parties to match haskell impl.
 - update serde
 - add json-diff for plutus data tests
- add indexmap for maps
- use plutus_data derive macro for contracts instead of custom impl
- refactor case enum to match haskell impl
- rename ada to lovelace in examples

+++ src/lib/deserialization/marlowe.rs
@@ -9,7 +9,7 @@ pub fn deserialize(content:&str) -> Result<ContractParseResult,ParseError>  {
     deserialize_with_input(content.trim_start().trim_end(),Default::default())
 }
 
-pub fn deserialize_with_input(content:&str,input:HashMap<String,i64>) -> Result<ContractParseResult,ParseError>  {
+pub fn deserialize_with_input(content:&str,input:HashMap<String,i128>) -> Result<ContractParseResult,ParseError>  {
     match <MarloweParser as pest::Parser::<Rule>>::parse(
         Rule::MainContract, 
         content
